requirements.txt
requests==2.32.3
pytest==8.3.5
python-dotenv
pytest-html